a lot

+++ src/main/java/com/buoobuoo/minecraftenhanced/MinecraftEnhanced.java
@@ -7,8 +7,6 @@ import com.buoobuoo.minecraftenhanced.core.block.CustomBlockManager;
 import com.buoobuoo.minecraftenhanced.core.chat.ChatManager;
 import com.buoobuoo.minecraftenhanced.core.damage.DamageManager;
 import com.buoobuoo.minecraftenhanced.core.entity.EntityManager;
-import com.buoobuoo.minecraftenhanced.core.entity.impl.ItemDropEntity;
-import com.buoobuoo.minecraftenhanced.core.entity.interf.CustomEntity;
 import com.buoobuoo.minecraftenhanced.core.event.DatabaseConnectEvent;
 import com.buoobuoo.minecraftenhanced.core.event.listener.*;
 import com.buoobuoo.minecraftenhanced.core.event.listener.mechanic.*;
@@ -18,7 +16,8 @@ import com.buoobuoo.minecraftenhanced.core.inventory.CustomInventoryManager;
 import com.buoobuoo.minecraftenhanced.core.item.CustomItemManager;
 import com.buoobuoo.minecraftenhanced.core.item.additional.attributes.ItemAttributeManager;
 import com.buoobuoo.minecraftenhanced.core.navigation.RouteManager;
-import com.buoobuoo.minecraftenhanced.core.party.PartyManager;
+import com.buoobuoo.minecraftenhanced.core.social.friends.FriendsManager;
+import com.buoobuoo.minecraftenhanced.core.social.party.PartyManager;
 import com.buoobuoo.minecraftenhanced.core.player.PlayerManager;
 import com.buoobuoo.minecraftenhanced.core.quest.QuestManager;
 import com.buoobuoo.minecraftenhanced.core.vfx.cinematic.SpectatorManager;
@@ -26,6 +25,8 @@ import com.buoobuoo.minecraftenhanced.permission.PermissionManager;
 import com.buoobuoo.minecraftenhanced.persistence.MongoHook;
 import com.comphenix.protocol.ProtocolLibrary;
 import com.comphenix.protocol.ProtocolManager;
+import com.mongodb.Mongo;
+import com.mongodb.MongoClientSettings;
 import lombok.Getter;
 import org.bukkit.Bukkit;
 import org.bukkit.GameRule;
@@ -36,14 +37,22 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerDropItemEvent;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.scheduler.BukkitRunnable;
 import org.bukkit.util.Vector;
+import org.slf4j.LoggerFactory;
+
+import java.lang.reflect.Method;
+import java.util.Properties;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 @Getter
 public class MinecraftEnhanced extends JavaPlugin implements Listener{
 
     public static final String MAIN_WORLD_NAME = "world";
+    public static final String BRAND_CHANNEL = "minecraft:brand";
 
     private MongoHook mongoHook;
 
@@ -63,6 +72,7 @@ public class MinecraftEnhanced extends JavaPlugin implements Listener{
     private ChatManager chatManager;
     private AreaManager areaManager;
     private RouteManager routeManager;
+    private FriendsManager friendsManager;
 
     private CommandManager commandManager;
 
@@ -72,6 +82,8 @@ public class MinecraftEnhanced extends JavaPlugin implements Listener{
 
     @Override
     public void onEnable(){
+        System.setProperty("logging.level.org.mongodb", "WARN");
+
         initManagers();
         initListeners();
         initTimers();
@@ -80,13 +92,25 @@ public class MinecraftEnhanced extends JavaPlugin implements Listener{
         for(World world : Bukkit.getWorlds()){
             world.setGameRule(GameRule.KEEP_INVENTORY, true);
         }
+        try {
+            Method registerMethod = this.getServer().getMessenger().getClass().getDeclaredMethod("addToOutgoing", Plugin.class, String.class);
+            registerMethod.setAccessible(true);
+            registerMethod.invoke(this.getServer().getMessenger(), this, MinecraftEnhanced.BRAND_CHANNEL);
+        } catch (ReflectiveOperationException e) {
+            throw new RuntimeException("Error while attempting to register plugin message channel", e);
+        }
     }
 
     @Override
     public void onDisable(){
-        entityManager.destroyAll();
-        playerManager.saveAll();
-        mongoHook.disable();
+        if(playerManager != null)
+            playerManager.saveAll();
+
+        if(mongoHook != null)
+            mongoHook.disable();
+
+        if (entityManager != null)
+            entityManager.destroyAll();
     }
 
     public void initManagers(){
@@ -108,6 +132,7 @@ public class MinecraftEnhanced extends JavaPlugin implements Listener{
         abilityManager = new AbilityManager(this);
         chatManager = new ChatManager(this);
         areaManager = new AreaManager(this);
+        friendsManager = new FriendsManager(this);
 
         commandManager = new CommandManager(this);
 
@@ -140,13 +165,16 @@ public class MinecraftEnhanced extends JavaPlugin implements Listener{
 
                 new ProjectileHitEventListener(this),
                 new ItemRequirementEventListener(this),
+                new InteractableBlockEventListener(this),
                 new BlockGrowEventListener(this),
                 new NpcTeamTagListener(this),
-                new MoistureChangeEventListener(this)
+                new MoistureChangeEventListener(this),
+                new ServerBrandListener(this)
 
         );
 
         //packet listener
+        new PlayerChatPacketListener(this);
         new PlayerInteractNpcPacketListener(this);
         new AnvilRenamePacketListener(this);
     }
@@ -167,6 +195,11 @@ public class MinecraftEnhanced extends JavaPlugin implements Listener{
 
         }.runTaskTimer(this, 0, 1);
 
+        Bukkit.getScheduler().runTaskLater(this, this::delayedInit, 20);
+    }
+
+    public void delayedInit(){
+        entityManager.initFixtures();
     }
 
     //TEMP
